Update student_test.dart (#27)
* Update student_test.dart

* Update student_notifier.dart

* What is this.

What even is this

+++ lib/notifiers/student_notifier.dart
@@ -43,7 +43,7 @@ class StudentNotifier extends UserNotifier{
   
 
   void setScores(int index, String scores) {
-    updatedTests[index].score = double.parse(scores) ;
+    updatedTests[index].score = int.parse(scores) ;
   }
 
   doneUpdating(){
